# src/main/java/com/example/datacenter/dto/rack/RackCreateRequest.java
package com.example.datacenter.dto.rack;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;

public record RackCreateRequest(
        @NotBlank String serialNumber,
        @NotBlank String name,
        String description,
        @Min(1) int spaceSize,
        @Min(1) int maxPowerUsage
) {}